+++ app/scripts/fhir/R4/lformsFHIR.js
@@ -106,6 +106,11 @@ fhir.SDC = _lforms_fhir_sdc_js__WEBPACK_IMPORTED_MODULE_1__["default"];
 Object(_lforms_fhir_sdc_converter_js__WEBPACK_IMPORTED_MODULE_2__["default"])(fhir.SDC);
 fhir.SDC.fhirVersion = fhirVersion; // Needed by lfData for fhirpath, etc.
 
+fhir.reservedVarNames = {};
+['context', 'resource'].forEach(function (name) {
+  fhir.reservedVarNames[name] = true;
+});
+
 /***/ }),
 /* 1 */
 /***/ (function(module, exports, __webpack_require__) {
@@ -19197,6 +19202,13 @@ var sdcExport = {
           }]
         }
       });
+    }
+
+    if (item._isHidden) {
+      targetItem.extension.push({
+        url: "http://hl7.org/fhir/StructureDefinition/questionnaire-hidden",
+        valueBoolean: true
+      });
     } // linkId
 
 
@@ -19561,7 +19573,7 @@ var sdcExport = {
    */
   _processResponseItem: function _processResponseItem(item, parentItem) {
     var targetItem = {};
-    var linkId = item._codePath; // if it is a section
+    var linkId = item.linkId ? item.linkId : item._codePath; // if it is a section
 
     if (item.dataType === "SECTION") {
       // linkId
@@ -19630,6 +19642,33 @@ var sdcExport = {
         "valueUri": item.externallyDefined
       });
     }
+  },
+
+  /**
+   * Make a FHIR Quantity for the given value and unit info.
+   * @param value required, must be an integer or decimal
+   * @param itemUnit optional, lform data item.unit (that has a name property)
+   * @param unitSystem optional, default to 'http://unitsofmeasure.org'
+   * @return a FHIR quantity or null IFF the given value is not a number (parseFloat() returns NaN).
+   * @private
+   */
+  _makeValueQuantity: function _makeValueQuantity(value, itemUnit, unitSystem) {
+    var fhirQuantity = null;
+    var floatValue = parseFloat(value);
+
+    if (!isNaN(floatValue)) {
+      fhirQuantity = {
+        value: floatValue
+      };
+
+      if (itemUnit && itemUnit.name) {
+        fhirQuantity.unit = itemUnit.name;
+        fhirQuantity.code = itemUnit.name;
+        fhirQuantity.system = unitSystem ? unitSystem : 'http://unitsofmeasure.org';
+      }
+    }
+
+    return fhirQuantity;
   },
 
   /**
@@ -19809,8 +19848,16 @@ var sdcExport = {
         //   "system" : "<uri>", // Code System that defines coded unit form
         //   "code" : "<code>" // Coded form of the unit
         // }]
-        else if (item.dataType === "QTY") {} // NOTE: QTY data type in LForms does not have unit. Cannot support it.
-          // make a Quantity type if numeric values has a unit value
+        else if (item.dataType === "QTY") {
+            // for now, handling only simple quantities without the comparators.
+            var fhirQuantity = this._makeValueQuantity(values[i], item.unit);
+
+            if (fhirQuantity) {
+              answer.push({
+                valueQuantity: fhirQuantity
+              });
+            }
+          } // make a Quantity type if numeric values has a unit value
           else if (item.unit && typeof values[i] !== 'undefined' && (item.dataType === "INT" || item.dataType === "REAL" || item.dataType === "ST")) {
               answer.push({
                 "valueQuantity": {
@@ -19906,8 +19953,14 @@ var sdcExport = {
       //   "system" : "<uri>", // Code System that defines coded unit form
       //   "code" : "<code>" // Coded form of the unit
       // }]
-      else if (item.dataType === 'QTY') {} // NOTE: QTY data type in LForms does not have unit. Cannot support it.
-        // for boolean, decimal, integer, date, dateTime, instant, time, string, uri
+      else if (item.dataType === 'QTY') {
+          // for now, handling only simple quantities without the comparators.
+          var fhirQuantity = this._makeValueQuantity(item.value, item.unit);
+
+          if (fhirQuantity) {
+            targetItem[valueKey] = fhirQuantity;
+          }
+        } // for boolean, decimal, integer, date, dateTime, instant, time, string, uri
         else if (item.dataType === "BL" || item.dataType === "REAL" || item.dataType === "INT" || item.dataType === "DT" || item.dataType === "DTM" || item.dataType === "TM" || item.dataType === "ST" || item.dataType === "TX" || item.dataType === "URL") {
             if (this._answerRepeats(item) && Array.isArray(item.defaultAnswer)) {
               for (var k = 0; k < item.defaultAnswer.length; k++) {
@@ -20001,8 +20054,14 @@ var sdcExport = {
         //   "system" : "<uri>", // Code System that defines coded unit form
         //   "code" : "<code>" // Coded form of the unit
         // }]
-        else if (sourceItem.dataType === 'QTY') {} // TBD
-          // for boolean, decimal, integer, date, dateTime, instant, time, string, uri
+        else if (sourceItem.dataType === 'QTY') {
+            // for now, handling only simple quantities without the comparators.
+            var fhirQuantity = this._makeValueQuantity(condition.trigger.value, sourceItem.unit);
+
+            if (fhirQuantity) {
+              enableWhenRule[valueKey] = fhirQuantity;
+            }
+          } // for boolean, decimal, integer, date, dateTime, instant, time, string, uri
           else if (sourceItem.dataType === "BL") {
               enableWhenRules[0].operator = 'exists'; // Spec says exists implies answer is boolean, then 'exists' is redundant, isn't it?
 
@@ -20350,9 +20409,12 @@ var sdcExport = {
         case "INT":
           if (qrValue.valueQuantity) {
             item.value = qrValue.valueQuantity.value;
-            item.unit = {
-              name: qrValue.valueQuantity.code
-            };
+
+            if (qrValue.valueQuantity.code) {
+              item.unit = {
+                name: qrValue.valueQuantity.code
+              };
+            }
           } else if (qrValue.valueInteger) {
             item.value = qrValue.valueInteger;
           }
@@ -20360,11 +20422,15 @@ var sdcExport = {
           break;
 
         case "REAL":
+        case "QTY":
           if (qrValue.valueQuantity) {
             item.value = qrValue.valueQuantity.value;
-            item.unit = {
-              name: qrValue.valueQuantity.code
-            };
+
+            if (qrValue.valueQuantity.code) {
+              item.unit = {
+                name: qrValue.valueQuantity.code
+              };
+            }
           } else if (qrValue.valueDecimal) {
             item.value = qrValue.valueDecimal;
           }
@@ -20448,7 +20514,11 @@ function addSDCImportFns(ns) {
   self.fhirExtUrlRestrictionArray = ["http://hl7.org/fhir/StructureDefinition/minValue", "http://hl7.org/fhir/StructureDefinition/maxValue", "http://hl7.org/fhir/StructureDefinition/minLength", "http://hl7.org/fhir/StructureDefinition/regex"];
   self.fhirExtUrlAnswerRepeats = "http://hl7.org/fhir/StructureDefinition/questionnaire-answerRepeats";
   self.fhirExtUrlExternallyDefined = "http://hl7.org/fhir/StructureDefinition/questionnaire-externallydefined";
+<<<<<<< HEAD
   self.argonautExtUrlExtensionScore = "http://fhir.org/guides/argonaut-questionnaire/StructureDefinition/extension-score";
+=======
+  self.fhirExtUrlHidden = "http://hl7.org/fhir/StructureDefinition/questionnaire-hidden";
+>>>>>>> master
   self.formLevelIgnoredFields = [// Resource
   'id', 'meta', 'implicitRules', 'language', // Domain Resource
   'text', 'contained', 'text', 'contained', 'extension', 'modifiedExtension', // Questionnaire
@@ -20547,6 +20617,8 @@ function addSDCImportFns(ns) {
     self._processRestrictions(targetItem, qItem);
 
     self._processCodingInstructions(targetItem, qItem);
+
+    self._processHiddenItem(targetItem, qItem);
 
     self._processUnitList(targetItem, qItem);
 
@@ -20730,6 +20802,25 @@ function addSDCImportFns(ns) {
     }
   };
   /**
+   * Parse questionnaire item for "hidden" extension
+   *
+   * @param lfItem {object} - LForms item object to be assigned the _isHidden flag if the item is to be hidden.
+   * @param qItem {object} - Questionnaire item object
+   * @private
+   * @return true if the item is hidden or if its ancestor is hidden, false otherwise
+   */
+
+
+  self._processHiddenItem = function (lfItem, qItem) {
+    var ci = LForms.Util.findObjectInArray(qItem.extension, 'url', self.fhirExtUrlHidden);
+
+    if (ci) {
+      lfItem._isHidden = typeof ci.valueBoolean === 'boolean' ? ci.valueBoolean : ci.valueBoolean === 'true';
+    }
+
+    return lfItem._isHidden;
+  };
+  /**
    * Parse questionnaire item for answers list
    *
    * @param lfItem {object} - LForms item object to assign answer list
@@ -20816,26 +20907,28 @@ function addSDCImportFns(ns) {
       var val = _getValueWithPrefixKey(elem, /^value/);
 
       if (lfItem.dataType === 'CWE' || lfItem.dataType === 'CNE') {
-        if (isMultiple) {
-          if (!defaultAnswer) defaultAnswer = [];
-          answer = {};
-          if (val.code !== undefined) answer.code = val.code;
-          if (val.display !== undefined) answer.text = val.display;
-          defaultAnswer.push(answer);
-        } // single selection
-        else {
-            answer = {};
-            if (val.code !== undefined) answer.code = val.code;
-            if (val.display !== undefined) answer.text = val.display;
-            defaultAnswer = answer;
-          }
-      } else {
-        if (isMultiple) {
-          if (!defaultAnswer) defaultAnswer = [];
-          defaultAnswer.push(val);
-        } else {
-          defaultAnswer = val;
+        answer = {};
+        if (val.code !== undefined) answer.code = val.code;
+        if (val.display !== undefined) answer.text = val.display;
+      } else if (lfItem.dataType === 'QTY') {
+        answer = val.value;
+        var unit = val.code ? val.code : val.unit;
+
+        if (unit) {
+          lfItem.unit = {
+            name: unit
+          };
         }
+      } else {
+        answer = val;
+      }
+
+      if (isMultiple) {
+        if (!defaultAnswer) defaultAnswer = [];
+        defaultAnswer.push(answer);
+      } else {
+        // single selection
+        defaultAnswer = answer;
       }
     });
     lfItem.value = defaultAnswer; // TODO - Is this necessary?