+++ app/scripts/fhir/STU3/lforms-fhir-sdc-converter.js
@@ -33,6 +33,7 @@ function addSDCImportFns(ns) {
   self.fhirExtUrlExternallyDefined = "http://hl7.org/fhir/StructureDefinition/questionnaire-externallydefined";
   self.argonautExtUrlExtensionScore = "http://fhir.org/guides/argonaut-questionnaire/StructureDefinition/extension-score";
 
+  self.fhirExtUrlHidden = "http://hl7.org/fhir/StructureDefinition/questionnaire-hidden";
 
   /**
    * Convert FHIR SQC Questionnaire to LForms definition
@@ -103,6 +104,7 @@ function addSDCImportFns(ns) {
     _processDisplayControl(targetItem, qItem);
     _processRestrictions(targetItem, qItem);
     _processCodingInstructions(targetItem, qItem);
+    _processHiddenItem(targetItem, qItem);
     _processUnitList(targetItem, qItem);
     self._processDefaultAnswer(targetItem, qItem);
     _processExternallyDefined(targetItem, qItem);
@@ -207,6 +209,23 @@ function addSDCImportFns(ns) {
 
 
   /**
+   * Parse questionnaire item for "hidden" extension
+   *
+   * @param lfItem {object} - LForms item object to be assigned the _isHidden flag if the item is to be hidden.
+   * @param qItem {object} - Questionnaire item object
+   * @private
+   * @return true if the item is hidden or if its ancestor is hidden, false otherwise
+   */
+  function _processHiddenItem(lfItem, qItem) {
+    var ci = LForms.Util.findObjectInArray(qItem.extension, 'url', self.fhirExtUrlHidden);
+    if(ci) {
+      lfItem._isHidden = typeof ci.valueBoolean === 'boolean'? ci.valueBoolean: ci.valueBoolean === 'true';
+    }
+    return lfItem._isHidden;
+  }
+
+
+  /**
    * Parse questionnaire item for answers list
    *
    * @param lfItem {object} - LForms item object to assign answer list
@@ -282,6 +301,16 @@ function addSDCImportFns(ns) {
         else {
           lfItem.value = {code: val.code, text: val.display};
           lfItem.defaultAnswer = {code: val.code, text: val.display};
+        }
+      }
+      else if(lfItem.dataType === 'QTY') {
+        if (val.value !== undefined) {
+          lfItem.value = val.value;
+          lfItem.defaultAnswer = val.value;
+        }
+        let unit = val.code? val.code: val.unit;
+        if (unit) {
+          lfItem.unit = {name: unit};
         }
       }
       else {