+++ test/protractor/spec/R4/prepop.spec.js
@@ -22,7 +22,7 @@ describe('Form pre-population', function() {
     });
     tp.loadFromTestData('ussg-fhp.json', 'R4');
   });
-  
+
   it('should be possible to pull in data from a FHIR context', function() {
     expect(tp.USSGFHTVertical.name.getAttribute('value')).toBe("John Smith");
     expect(tp.USSGFHTVertical.dob.getAttribute('value')).toBe("12/10/1990");